Moved resources into their own version folder to increase flexibility of releases later on
We want to be able to completely rewrite css and js sometime down the line without affecting current implementations. To complete this we are now moving resources into its own version folder. i.e. 14.2.0...if we do a major rewrite, we can simply do, 14.2.1 and make any new implementation of views point to that.