[FUNNELAPPS-226] Refactor Groovy hook script into classes
Now that we can use custom Groovy classes in hook scripts (FUN-7664) we
can refactor Hook script logic into separate classes.

I created a basic class structure where `StencilHooks` is the main entry
point that will need to be manually called in the collection hook
scripts. As suggested by @gtran It will then run the scripts for all the
stencils, for the relevant hook phase (pre process, post process, etc.).

`TabsHookLifecycle` is typically what stencils developers will have to
implement for their stencils. When a new stencil-specific class is
created it will also need to be added to the list of classes in
`StencilHooks.HOOKS`. That's all that should be needed, the core logic
doesn't need to be modified or even understood by the developer.

`TabsHookLifecycle` is just an example for now and doesn't do anything.

Each stencil hook class will be responsible for checking if it should
run or not (probably based on `collection.cfg` settings).

I also wanted all that to be unit testable so I added a Gradle build
script and some example tests. This will also allow us to integrate this
with GitLab CI so that each merge requests will run the unit test and
notify the MR if the build fails.

I initially went with Maven but the Groovy support is sub par, whereas
Gradle seem to work well. It's also supported by most IDEs which makes
importing the project possible.

*Warning*: I had to move the source files around so that the package
declarations match the file paths, in order for IDE support to work
correctly. We'll have to make sure that when this is deployed the
`stencils` symlink from `lib/java/groovy/` is changed to point to
`src/main/groovy/com/funnelback/stencils/`.