Implement Groovy hook script for the Tabs stencil
Converted the code that was in the Tabs post datafetch hook script into
a "proper" class as per the new code structure + added corresponding
unit test

Aside: Changed the logic of the hook helper a bit to stop storing the
`transaction` as an instance variable, but rather pass it to each hook
method. That's safer as the hook instance (From the `HOOKS` constant
array) will actually be shared between all the requests so it must not
hold any state.