Cleanup README and add install instructions
I had to install the Stencils again on my laptop and I couldn't remember
how, and having to go to Confluence is a pain so I put some minimal
instructions in there.

Also added the context file for Jetty so that we don't have to download
it from Confluence, and we can update it in GitLab if needed.