Move controller logic inside the data model for Facebook
As discussed with @gtran, attempt to get rid of the controller tags in
favor of augmenting the existing data model, so that the Stencils
FreeMarker tags can be view / presentation only.

Added a hook for Facebook which re-implements what was done in
the Facebook controller but in a hook script and place the relevant data
inside each result `customData` map.

In addition, implemented some custom FreeMarker methods that can be
placed in the data model and used inside the template with
`${response.customData.stencilsMethods.methodName(...)}`.

Took the existing Facebook view, and converted it into a simpler macro
library that makes use of the new datamodel. Also removed implicit use
of variables on the `core` namespace, I've replaced them by explicit
parameter so that parent tags have to pass in the `result` they will
use.  Making things more explicit will make them easier to follow, less
magic and less error prone (Note that it will need changes to the core /
base controllers to allow explicit parameters to be passed).

To avoid disruption on existing stencils, moved this updated view into
a different file `libraries/facebook/facebook.ftl`. A single file is enough
since there's no controllers anymore. This file can be imported with a
regular `<#import>` statement (no more magic with the stencil utilities).

Benefits:
- No more FreeMarker implementing non-presentation logic which is often
hard to follow
- It can be unit tested
- It's a lot more discoverable to have the data in the data model that
can be viewed easily with the JSON endpoint, rather than having to infer
what variables the FreeMarker tags may put in the scope
- It also helps users integrating directly with the JSON/XML endpoint
- It can be more easily retro-fitted in the product (just convert the
Groovy code into Java, a bit like when we productise Groovy filters)

I also added a hack to inject the current query string into the data
model as a map, because it will be needed to construct URLs to the
search (e.g.  explore URL for each result, URL for facet categories,
etc.). These URLs are constructed usually by taking the "current" URL
and adding/removing parameters.

In 15.10 the data model will include the query string as a map inside
`question`. It already exists actually but is not in the `transaction`,
so it's unavailable to hook scripts.

This map is immutable so that developers cannot alter the query string.
There's a utility method to make a copy of it, to be used when building
URLs to easily add/update/remove parameters.