Avoid a NullPointerException if the query string map is missing
This will happen with extra searches, as they run in a different thread
we don't have access to the original request to get the query string
from. This is actually common when using the Tab Stencil as it uses an
extra search.

When we don't have a source map to copy from, just return an empty map.
That will allow the hooks to work, albeit in degraded mode where the
generated URLs will miss some parameters.