[FUNNELAPPS-229] Added generic script to generate CSV completions
Added script to generate per-profile CSV completions, assuming a
template exists in the profile with the name `auto-completions.ftl`.
This script will query the profile on the given view, save the CSV data
and call `build_autoc` over it.

I took inspiration from existing scripts + the one @lnowak wrote.

That should allow implementations (Stencils and others) to just call
this script directly with the right parameters rather than having to
re-implement it over and over again.