Make the query string map available for extra searches
It didn't work before because we only attempted to populate the extra
searches if the main transaction didn't already have the query string map
already.

In practice the main transaction may get the query string map as part of
pre_process, pre_datafetch or post_datafetch hooks. At this time the
extra searches are not available so they don't receive the query string
map.

By the time we reach post_process where the extra searches are
available we don't attempt to inject the query string map since the map
is already in the main transaction.

Changed the code to always attempt to inject the map in extra searches,
if we have a query string map.

This is required to use the new facets data model in conjunction with
the tabs stencils, as they rely on the FACETED_NAVIGATION extra search.