Update documentation
Remove placeholder MD files and provide actual documentation for the
hook scripts, filters and workflow scripts.

I kept the documentation sparse and focused on the setup / configuration
on purpose to make it easier to read and quickly get started. Hopefully
the code is readable / commented enough so that users can look into it
if they want more details.
In addition with the new approach looking at the generated data model is
probably the easiest way to see what's going on.