Inject a fake "All" value in the Tabs facet
Inject a fake "All" value for the "all results" tab in the "Tabs"
Stencil facet, if present. This value has URLs that removes any tab
facet constraint, to return the full data set.

This makes it unnecessary to have a real "all" value mapped to a gscope
or a metadata, which was cumbersome as we needed to make sure absolutely
all content was tagged with the correct gscope or metadata. This is
difficult to guarantee on large crawls where a lot of domain names are
not known in advance.

In addition having this fake value in the data model means the
FreeMarker code doesn't have to worry about providing a fake "All" link,
it's just treated as a regular tab.

The fake "All" will be `selected=true` when no other Tabs are.