Store the query string map in the question rather than the response
The query string map is used by various hooks to compute URLs derived
from the "current" URL (e.g.  facet URLs, explore URL, etc.). It's
available in 15.10, but not in 15.8 or earlier. To cater for this, custom
code was written to build a map from the current query string and store
it in the response `customData`.

This was however problematic for extra searches, because the current
query string cannot be retrieved for extra searches (because they run in
a separate thread). There was no way to inject a query string map in the
extra searches because:
- If we do it in `pre_*` hooks there's no `response` object yet to store
it in
- If we do it in `post_*` hooks the extra searches have already
completely run so it's too late

However we need this to work with extra searches for the Tabs stencils
where the tab values are read from the FACETED_NAVIGATION extra search.

To make that work, I added a `customData` map in the search question in
the Modern UI. This way we can store the query string map in the
question and it will be made available to extra searches automatically
because they are cloned from the main search. I backported that to 15.8
and 15.10 in a patch.

This MR changes the hooks to account for that, store the query string
map in the question and read it back when needed in the individual
hooks.

I also had to cater for different Funnelback versions and patch levels,
depending if the `question.customData` is present or not. A bit ugly,
but that's how backwards compatibility is...