Refinement to the Tab facet
Currently the tab values and counts are read from the FACETED_NAVIGATION
extra search, but the current tab being selected is read from the main
search. That's because the extra search has no facet constraints on
purpose as we want full counts, so only the main search has a "current
tab" constraint.

That makes it difficult to display the tabs in FreeMarker because the
counts and labels must be taken from the extra search, but the "active"
status need to be read from the main search. To mitigate this, update
the tabs hook to copy the "selected" flag from the main search into the
extra search.  This way FreeMarker can work with the extra search only
and not have to worry about the main search.

Also updated the tab `selectUrl` for each tab to remove any other facet
constraint.  That's needed to:
* Ensure that we drop any tab-specific facet constraint when switching
tab. For example on the "Courses" tab there may be a currently selected
facet "Type = On Campus". If that was preserved while we switched to the
"People" tab that would result in zero result as no people results have
a metadata "Type = On Campus"
* Work with Gscope based tabs: Gscope based tabs have different query
string parameter names (`f.Tab|1=Courses`, `f.Tab|2=People`). If we
don't remove all the facet values in the URL, then multiple tabs may be
selected at once, resulting in zero results
(`&f.Tab|1=Courses&f.Tab|2=People`).