[FUN-10375] Remove the "site" cluster if there's only one site
Displaying a site cluster with one site is useless because clicking on
the cluster will have no effect (as all the results are already coming
from the suggested site).

This single-site cluster is in fact hidden in FreeMarker but that's
cumbersome to do as FreeMarker needs to do a deep inspection of the data
model to decide if the contextual navigation panel should be displayed
or not (as it should not be displayed if there's only 1 cluster, of type
site, and that this cluster contains only 1 category).

To make that easier, just update the data model to remove the site
cluster when it only has one site, so that we don't need special cases
in FreeMarker.