Select the result template based on the result gscope
In some cases we want results to look different within the same
collection, so selecting the template based on the result source
collection is not enough. Added support to it via the result gscopes,
looking up if a template has been configured for each gscope number the
result has.

Aside: Fix indentation of tabs macro