Fix handlers for more/less buttons on facets
We can't select the UL element, because then when we're trying to get
back to it from the more/less button jQuery can't find it (via
`closest()`) because the button is a sibling of the UL, not a children.

Instead just use a selector for the whole facet DIV, rather than the UL.
This should cater for any changes in the button HTML, as long as it's
contained inside the facet DIV (which make sense).