Remove the Tabs facet from the list of selected facets
The Tab facet is special and we don't want it to appear in the list of
currently selected facets ("Refined by: ..."). To avoid having to deal
with that in FreeMarker, update the Tabs hook to remove the tabs facet
from the list of selected facets.

I had to change the type for the node that contains the list of selected
facets from array to list, as it makes it easier to remove a single
value (it doesn't impact FreeMarker or Groovy scripts as they both treat
lists and arrays the same).

The currently selected tab was already put in a separate data model node
in case it was needed in the front-end