Ensure the "All" tab reset all facets constraints
All the other tabs reset the facets constraints (so that when switching
tabs we don't preserve a facet constraint that doesn't make sense, such
as "Position = Professor" when switching to the "Events" facet). The
"All" tab wasn't doing that though because it's a fake facet value that
was injected *after* we were modifying the facet URLs.

To fix that, ensure we inject the fake "All" facet value first, and then
update all the URLs to remove the selected facets.