Support different shortlist templates per result type
Allow different shortlist item types (based on their source collection) to have
different shortlist templates, using a mechanism similar to what is done
to automatically select the right result template in the main search.

With that, a "course" can look different than a "staff" in the
shortlist. Each template is held by the relevant FreeMarker library
(e.g. `course.ftl`, `staff.ftl`), with the idea that the shortlist macro
will live next to the search result macro to make sure they stay in
sync.

Aside: Simplified a bit the template selection system, since I found
that you can use the special variable `.main` to access the root
namespace. This way the caller doesn't have to pass it in which looked a
bit like magic in the main `project.ftl` template.