Added script to partialize CSV auto-completions
`build_autoc`'s `-partials` only works on organic suggestions, and
having partial support for CSV completions is often desirable.

Added a script that takes an input CSV file and generate an output CSV
file with the trigger column "partialized" in a similar fashion as
`build_autoc`. An optional stop words file can be passed in.

Aside: Remove extra class path when running the workflow commands. It's
not needed as when the Stencils are installed a symlink is created
inside `$SEARCH_HOME/lib/java/groovy/`, putting the Stencils classes on
the classpath.