Avoid relying on `$SEARCH_HOME` and a specific Groovy version
This improves the build script a bit to not rely on a specific
`$SEARCH_HOME` nor Groovy version. This is especially useful when
running 2 different version of Funnelback on the same machine.

Additionally it means that we don't have to update the Groovy version
for each new Funnelback release.

The GitLab CI config had to be modified accordingly because it was not
checking out the code under `$SEARCH_HOME/share/stencils/`.