Ensure phrase are still matched with the CSV partializer
Previously we were splitting the original trigger words into individual
words and adding trigger for each of them. That would not preserve the
original phrase that would stop being matched from partial queries.

For example with the trigger "chocolate cake", we would generate a CSV
row for "chocolate" and "cake", but it would not match the partial query
"chocolate ca" for example.

Fix this by just copying the original row over if it's a phrase (more
than one word), so that the initial behaviour will be preserved.