Fix node containing the selected tab for 15.12
The Stencils populate a node containing the currently selected tab in
the data model. By default the "All" tab is selected (i.e. no other tabs
are selected). If we detect Stencils-type selected facets in the data
model we then iterate through these and replace it with the actual
selected tab (e.g.  "Courses")

If we don't find any selected tab and are on 15.12, we try to get the
selected tab from the new 15.12 data model (as there's now a built-in
Tab-style facet, with a built-in "All" value when no other tab are
selected).

There was an edge case where you may be using 15.12 with legacy-style
facet, in which case the data model doesn't have any "Tab" style facets
and we can't locate the selected tab from the new 15.12 data model.

Since we couldn't locate the selected tab the "All" facet was completely
removed from the data model, and there was no more "default" tab
selected.

To fix this, do not remove the default "All" selected tab by default. If
we're using legacy facets, that will select "All" by default. If we're not
using legacy facets then the active tab will be found in the 15.12 data
model and overwrite the default "All" selection in the data model.