[PS-65] Fix the sort and limit dropdown
The dropdowns are providing links with pre-canned `sort` and `num_ranks`
parameters, but are not removing existing parameters in the current URL,
resulting in duplicated parameters.

Use the `removeParam()` function to remove the `sort` and `num_ranks`
parameter from the URL before building the new URL with the desired
sorting and limit values.