Added a generic CSV auto-completion template
Re-use Pete's idea of configuring in `collection.cfg` what triggers
should be used, as well as using the stopwords list.

Everything is done directly in the template without the need for hook
scripts which makes the setup a bit easier.

Collections wanting to use this template just have to `#include` it in
their own auto-completion template.