Update CSV generator to paginate through results
The previous version was just requesting all the results in one go,
which is fine for small collections but crashes for larger ones. I knew
that day would come when I did the initial naive implementation, but now
I need a smarter one for IRMI (~80,000 docs).

Hit the search results endpoints as many time as necessary to get the
data, varying `start_rank` for each request. Stop when we don't get any
data back anymore (meaning we've exceeded the number of docs in the
collection).