Fix bug in CSV completion generator
The latest change made the CSV completion generator append to the CSV
file (as we are paging through results) rather than write the file in
one go.

As a result if the CSV file pre-existed we would continue appending to
it indefinitely at each collection update. To fix this, delete the file
before generating it again.