Deal with `Q` action type for auto-complete
When the action is set to `Q` a search will be run when the user selects
a suggestion. The keywords that are used come from the "action" column,
which was set to the URL by default, trigger a search on the URL which
is not desirable.

Change it so that when the action type is `Q`, the result title is used
for searching instead. We may need to make that configurable at a later
stage.