Added a helper class to gather Instagram posts
This supersedes the dodgy Instagram connector that we've been using so
far. Benefits:
- Lot simpler to setup: No more JAR to install, reads settings from
`collection.cfg`, integrates with regular custom collections (no need
for custom workflows)
- Can be used from any collection by referencing the Stencils without
having to copy code in
- Integrates better with Funnelback, in terms of logging and progress
report when the update is running

It basically does the same thing, but better and simpler. It still
doesn't deal with authentication as Instagram has no API to
programmatically authenticate.