Updated GitLab CI configuration to synchronize with a public GitHub repo
Added configuration to synchronize the GitLab branches with a public
GitHub repo, to make the Stencils publicly available to partners and
clients.

This is done via a SSH deploy key that has been configured on GitHub and
is allowed to push to the GitHub repository.

On the GitLab side we store the key in a GitLab CI variable (in the
projects settings). This variable is flagged "protected" so that it will
only be exposed to protected branches (master, and the release
branches).  That way we only sync these branches to GitHub, and not
every single dev/feature branch that's created.

With that setup, anything merged to master or any of the release
branches will automatically be pushed to GitHub for public availability.