Updated CI config to build and test with all supported Funnelback versions
The 15.12+ branch of the Stencils is supposed to work with 15.12 and
above (15.14, 15.16) so update the CI config to run the tests with all
these environments.

This way if we build something that doesn't work on a more recent
version we'll be alerted, and we will be able to decide to either make
it forward compatible or to just create a new branch for the problematic
version.

Also changed the Docker images: Stored them into the project Docker
registry rather than rely on the R&D Docker server. That way we can use
any available CI runner rather than being restricted to the R&D one.
Builds will be faster as a result, especially since now we run 3 builds
in parallel (15.12, 15.14, 15.16).