Force cloning of the repository for the GitHub sync job
The default `GIT_STRATEGY` is `fetch`, meaning the `.git/` folder of the
project will be kept across builds and GitLab CI will just fetch
changes.

This may cause issues especially when we revert changes or force push to
the `master` branch as the local `.git/` folder gets out of sync and
prevents pushing to GitHub.

To avoid this, force a `clone` strategy just for the part of the build
that pushes to GitHub.