FUNNELAPPS-455 - Changed the release version as we are now targeting 15.24 instead of 15.20
Initially, the changes to the base.js were to be released to version 15.20 of Funnelback. However, the team decided that new features should not be pushed out to existing implementations. Therefore, we are now going to target version 15.24 instead. We need to rename the base release folder to reflect this.